Time: 0 ms (100.00%) | Memory: 7.9 MB (37.42%) - LeetSync

+++ 3918-check-divisibility-by-digit-sum-and-product/check-divisibility-by-digit-sum-and-product.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool checkDivisibility(int n) {
-        long long sum=0, prod=1;
+        int sum=0, prod=1;
         int temp = n;
         int d = 0;
         while(temp > 0)
